Double to float fixes.

+++ ArduMower/Battery.cpp
@@ -37,26 +37,26 @@ void battery_setup(const uint8_t pinVoltage,
   ADCMan.setCapture(battery_p->pinVoltage, 1, false);
   ADCMan.setCapture(battery_p->pinChargeVoltage, 1, false);
 
-  battery_p->monitored = false;                // Monitor battery and charge voltage?
-  battery_p->batGoHomeIfBelow = 11.8;            // Drive home voltage (Volt)
-  battery_p->batSwitchOffIfBelow = 10.8;         // Switch off if below voltage (Volt)
-  battery_p->batSwitchOffIfIdle = 1;               // Switch off battery if idle for x minutes
-  battery_p->batFactor = 0.495;                  // Battery conversion factor
-  battery_p->batChgFactor = 0.495;               // Battery conversion factor
-  battery_p->batFull = 14.7;                     // Battery reference Voltage (fully charged)
-  battery_p->batChargingCurrentMax = 1.6;        // Maximum current your charger can deliver
-  battery_p->batFullCurrent = 0.3;               // Current flowing when battery is fully charged
-  battery_p->startChargingIfBelow = 13.5;        // Start charging if battery Voltage is below
-  battery_p->chargingTimeout = 12600000; // Cafety timer for charging (ms) 12600000 = 3.5hrs
+  battery_p->monitored = false;                  // Monitor battery and charge voltage?
+  battery_p->batGoHomeIfBelow = 11.8f;           // Drive home voltage (Volt)
+  battery_p->batSwitchOffIfBelow = 10.8f;        // Switch off if below voltage (Volt)
+  battery_p->batSwitchOffIfIdle = 1;             // Switch off battery if idle for x minutes
+  battery_p->batFactor = 0.495f;                 // Battery conversion factor
+  battery_p->batChgFactor = 0.495f;              // Battery conversion factor
+  battery_p->batFull = 14.7f;                    // Battery reference Voltage (fully charged)
+  battery_p->batChargingCurrentMax = 1.6f;       // Maximum current your charger can deliver
+  battery_p->batFullCurrent = 0.3f;              // Current flowing when battery is fully charged
+  battery_p->startChargingIfBelow = 13.5f;       // Start charging if battery Voltage is below
+  battery_p->chargingTimeout = 12600000;         // Safety timer for charging (ms) 12600000 = 3.5hrs
   battery_p->chgSenseZero = 511;                 // Charge current sense zero point //TODO: autocalibrate?
   battery_p->chgFactor = 39;                     // Charge current conversion factor
-  battery_p->chgSense = 185.0;                   // Sensitivity of the charging current sensor (mV/A) (For ACS712 5A = 185)
-  battery_p->chgChange = 0;                       // Reading reversal from - to + 1 or 0
+  battery_p->chgSense = 185.0f;                  // Sensitivity of the charging current sensor (mV/A) (For ACS712 5A = 185)
+  battery_p->chgChange = 0;                      // Reading reversal from - to + 1 or 0
   // sensor output console      (chgSelection = 0)
   // settings for ACS712 5A     (chgSelection = 1 / chgSenseZero ~ 511 / chgFactor = 39 / chgSense = 185.0 / chgChange = 0 oder 1 (je nach Stromrichtung) / chgNull = 2)
   // settings for INA169 board  (chgSelection = 2)
-  battery_p->chgSelection = 2;                    // Senor selection
-  battery_p->chgNull = 2;                          // Zero crossing charge current sensor
+  battery_p->chgSelection = 2;                   // Sensor selection
+  battery_p->chgNull = 2;                        // Zero crossing charge current sensor
 }
 
 void battery_read(Battery* battery_p)
@@ -65,18 +65,18 @@ void battery_read(Battery* battery_p)
   if (abs(battery_p->chgCurrent) > 0.04 && battery_p->chgVoltage > 5)
   {
     // charging
-    battery_p->batCapacity += (battery_p->chgCurrent / 36.0); //TODO: What is 36.0?
+    battery_p->batCapacity += (battery_p->chgCurrent / 36.0f); //TODO: What is 36.0?
   }
 
-  // convert to double
+  // convert to float
   int16_t batADC = ADCMan.read(battery_p->pinVoltage);
-  double batvolt = (double)batADC * battery_p->batFactor;
+  float batvolt = (float)batADC * battery_p->batFactor;
   int chgADC = ADCMan.read(battery_p->pinChargeVoltage);
-  double chgvolt = (double)chgADC * battery_p->batChgFactor;
-  double current = (double)ADCMan.read(battery_p->pinChargeCurrent);
+  float chgvolt = (float)chgADC * battery_p->batChgFactor;
+  float current = (float)ADCMan.read(battery_p->pinChargeCurrent);
 
   // low-pass filter
-  const double accel = 0.01;
+  const float accel = 0.01f;
 
   if (abs(battery_p->voltage - batvolt) > 5)
   {
@@ -84,7 +84,7 @@ void battery_read(Battery* battery_p)
   }
   else
   {
-    battery_p->voltage = (1.0 - accel) * battery_p->voltage + accel * batvolt;
+    battery_p->voltage = (1.0f - accel) * battery_p->voltage + accel * batvolt;
   }
 
   if (abs(battery_p->chgVoltage - chgvolt) > 5)
@@ -93,7 +93,7 @@ void battery_read(Battery* battery_p)
   }
   else
   {
-    battery_p->chgVoltage = (1.0 - accel) * battery_p->chgVoltage + accel * chgvolt;
+    battery_p->chgVoltage = (1.0f - accel) * battery_p->chgVoltage + accel * chgvolt;
   }
 
   //Deaktiviert für Ladestromsensor berechnung
@@ -116,10 +116,10 @@ void battery_read(Battery* battery_p)
   if (battery_p->chgSelection == 1)
   {
     float chgAMP = current;              //Sensorwert übergabe vom Ladestrompin
-    float vcc = 3.30 / battery_p->chgSenseZero * 1023.0; // Versorgungsspannung ermitteln!  chgSenseZero=511  ->Die Genauigkeit kann erhöt werden wenn der 3.3V Pin an ein Analogen Pin eingelesen wird. Dann ist vcc = (float) 3.30 / analogRead(X) * 1023.0;
-    float asensor = chgAMP * vcc / 1023.0;              // Messwert auslesen
-    asensor = asensor - (vcc / battery_p->chgNull); // Nulldurchgang (vcc/2) abziehen
-    battery_p->chgSense = battery_p->chgSense - ((5.00 - vcc) * battery_p->chgFactor); // Korrekturfactor für Vcc!  chgFactor=39
+    float vcc = 3.30f / battery_p->chgSenseZero * 1023.0f; // Versorgungsspannung ermitteln!  chgSenseZero=511  ->Die Genauigkeit kann erhöt werden wenn der 3.3V Pin an ein Analogen Pin eingelesen wird. Dann ist vcc = (float) 3.30 / analogRead(X) * 1023.0;
+    float asensor = chgAMP * vcc / 1023.0f;              // Messwert auslesen
+    asensor = asensor - (vcc / (float)battery_p->chgNull); // Nulldurchgang (vcc/2) abziehen
+    battery_p->chgSense = battery_p->chgSense - ((5.00f - vcc) * battery_p->chgFactor); // Korrekturfactor für Vcc!  chgFactor=39
     float amp = asensor / battery_p->chgSense * 1000;               // Ampere berechnen
     if (battery_p->chgChange == 1)
     {
@@ -140,7 +140,7 @@ void battery_read(Battery* battery_p)
   {
     float chgAMP = current;
     float asensor = (chgAMP * 5) / 1023; // umrechnen von messwert in Spannung (5V Reference)
-    float amp = asensor / (10 * 0.1); // Ampere berechnen RL = 10k    Is = (Vout x 1k) / (RS x RL)
+    float amp = asensor / (10 * 0.1f); // Ampere berechnen RL = 10k    Is = (Vout x 1k) / (RS x RL)
     if (amp < 0.0)
     {
       battery_p->chgCurrent = 0.0;
